Restructure task planning classes in preparation for PDDL processing
