--- src/robotics_utils/abstractions/pddl/__init__.py
@@ -1,4 +1,0 @@
-"""Import PDDL-related classes and definitions."""
-
-from .pddl_domain import PDDLDomain as PDDLDomain
-from .pddl_problem import PDDLProblem as PDDLProblem
