+++ src/robotics_utils/planning/task_planning/ground_atom.py
@@ -1,4 +1,4 @@
-"""Define classes to represent ground atoms (AKA grounded predicates) and their negations."""
+"""Define a class to represent ground atoms (AKA grounded predicates)."""
 
 from __future__ import annotations
 
@@ -6,8 +6,8 @@ from dataclasses import dataclass
 from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
-    from robotics_utils.abstractions.symbols.objects import ObjectSymbol
-    from robotics_utils.abstractions.symbols.predicate import Predicate
+    from robotics_utils.planning.task_planning.object_symbols import ObjectSymbol
+    from robotics_utils.planning.task_planning.predicate import Predicate
 
 
 @dataclass(frozen=True)
@@ -18,14 +18,14 @@ class GroundAtom:
     """The predicate that was grounded to create this ground atom."""
 
     arguments: tuple[ObjectSymbol, ...]
-    """Objects bound to the predicate's parameters (their order matches the parameters)."""
+    """Objects bound to the predicate's parameters (order matches the parameters)."""
 
     def _key(self) -> tuple:
         """Define a hash key to uniquely identify the ground atom."""
         return (self.predicate._key(), self.arguments)  # noqa: SLF001
 
     def __eq__(self, other: object) -> bool:
-        """Evaluate whether this ground atom and another are equal."""
+        """Evaluate whether this ground atom is equal to another."""
         if not isinstance(other, GroundAtom):
             return NotImplemented
 
@@ -43,12 +43,3 @@ class GroundAtom:
     def name(self) -> str:
         """Retrieve the name of the grounded predicate."""
         return self.predicate.name
-
-
-@dataclass(frozen=True)
-class GroundLiteral:
-    """A ground literal is a ground atom or its negation."""
-
-    ground_atom: GroundAtom
-    negated: bool
-    """True if the ground literal is negated, else False."""
